requirements.txt
ffmpeg-python
ollama
openai
python-dotenv
